MANTA-4988 Want mlocate to work with buckets
