+++ lib/common.js
@@ -8,6 +8,7 @@
  * Copyright 2020 Joyent, Inc.
  */
 
+var crypto = require('crypto');
 var EventEmitter = require('events').EventEmitter;
 var http = require('http');
 var once = require('once');
@@ -17,6 +18,7 @@ var util = require('util');
 var httpSignature = require('http-signature');
 
 var assert = require('assert-plus');
+var bignum = require('bignum');
 var vasync = require('vasync');
 var restifyErrors = require('restify-errors');
 var VError = require('verror');
@@ -792,6 +794,43 @@ function getSharkInfo(req, hostname) {
     return (sharks[0]);
 }
 
+// Maps a key to the location of its metadata in Manta.
+//
+// This function is in common.js so the mlocate tool can use it as well.
+//
+// Input:
+//      placementData: A placementData object retrieved from buckets-mdplacemet
+//      tkey: the key to locate
+// Output:
+//      an object with the location data, of the form:
+//      {
+//          pnode (string)
+//          vnode (integer)
+//          data (integer)
+//      }
+function getDataLocation(placementData, tkey) {
+    assert.object(placementData, 'placementData');
+    assert.string(tkey, 'tkey');
+
+    var value = crypto.createHash(placementData.ring.algorithm.NAME).
+        update(tkey).digest('hex');
+    // find the node that corresponds to this hash.
+    var vnodeHashInterval =
+        placementData.ring.algorithm.VNODE_HASH_INTERVAL;
+
+    var vnode = parseInt(bignum(value, 16).div(bignum(vnodeHashInterval, 16)),
+        10);
+
+    var pnode = placementData.ring.vnodeToPnodeMap[vnode].pnode;
+    var data = placementData.ring.pnodeToVnodeMap[pnode][vnode];
+
+    return {
+        pnode: pnode,
+        vnode: vnode,
+        data: data
+    };
+}
+
 ///--- Exports
 
 module.exports = {
@@ -947,5 +986,7 @@ module.exports = {
     findSharks: findSharks,
     startSharkStreams: startSharkStreams,
     sharkStreams: sharkStreams,
-    streamFromSharks: streamFromSharks
+    streamFromSharks: streamFromSharks,
+
+    getDataLocation: getDataLocation
 };
