+++ lib/metadata_placement.js
@@ -5,7 +5,7 @@
  */
 
 /*
- * Copyright 2019 Joyent, Inc.
+ * Copyright 2020 Joyent, Inc.
  */
 
 /*
@@ -17,9 +17,7 @@
 var EventEmitter = require('events').EventEmitter;
 
 var assert = require('assert-plus');
-var bignum = require('bignum');
 var buckets_mdapi = require('buckets-mdapi');
-var crypto = require('crypto');
 var fast = require('fast');
 var jsprim = require('jsprim');
 var once = require('once');
@@ -28,6 +26,7 @@ var vasync = require('vasync');
 var verror = require('verror');
 
 var buckets_mdapi_client = require('./buckets_mdapi_client');
+var common = require('./common');
 
 ///--- Globals
 
@@ -228,23 +227,7 @@ MetadataPlacement.prototype._getLocation = function _getLocation(tkey) {
         tkey: tkey
     }, 'getLocation: entered');
 
-    var value = crypto.createHash(this.placementData.ring.algorithm.NAME).
-        update(tkey).digest('hex');
-    // find the node that corresponds to this hash.
-    var vnodeHashInterval =
-        this.placementData.ring.algorithm.VNODE_HASH_INTERVAL;
-
-    var vnode = parseInt(bignum(value, 16).div(bignum(vnodeHashInterval, 16)),
-        10);
-
-    var pnode = this.placementData.ring.vnodeToPnodeMap[vnode].pnode;
-    var data = this.placementData.ring.pnodeToVnodeMap[pnode][vnode];
-
-    var ret = {
-        pnode: pnode,
-        vnode: vnode,
-        data: data
-    };
+    var ret = common.getDataLocation(this.placementData, tkey);
 
     log.debug({
         data: ret
